CAS: Switch from SHA1 to BLAKE3
Switch from SHA1 to BLAKE3 for the builtin CAS.

Really we should be adding an enum to choose, but that's kind of hard.
For now, I've reduced the places that name the hash down to two -- a
typedef and BuiltinCAS::getHashName() -- and added some FIXMEs to both
locations that sketch what it could look like.

Why is it hard? Mainly because we'd need to deal with error paths, such
as opening a valid builtin CAS that had a different hash function than
the new/current configuration. Not insurmountable, but let's just move
over.

Why is it important? I left a big FIXME in BuiltinCAS.h, but the short
version is: future-proofing (for moving to a new hash function) and
storage optimization (using truncated hashes).

This bumps the FilePrefix for the OnDiskCAS from "v1." to "v2." to allow
SHA1 and BLAKE3 objects to coexist at the same CAS path. Without this
change, I can't build locally! This is because my host toolchain is
CAS-enabled (and uses SHA1), and I have `-fdepscan` on for both the host
clang and for tablegen (which uses BLAKE3).