[lldb] When matching entry values for async swift function, fallback to the current function from parent function when evaluating subexpressions.
This occurs when the parent function does not have debug info (*). In swift async
functions this is ok since we only need to be able to restore the ABI defined
context register for the async context (r14)... so it is not necessary to match
with a callsite parameter.

(*) E.x.: On Linux async funclets called from a libswift_Concurrency.so that had its
debug info stripped.