CAS: Update description for OnDiskCAS layout
Update the description for the layout of the OnDiskCAS on disk. It was
originally written before being implemented, and had bitrotted as
files and tables were renamed. The new description:

- Refers to helper functions for table names, so that they're less
  likely to bitrot (since when the helper functions are renamed a local
  file search should catch this too).
- Describes missing pieces that didn't exist earlier (such as the
  actions table and standalone blobs).
- More precisely describes the contents, pointing at other helper
  data structures.

In response to review of 6ddc4144461c0d1571069d79561e733b7311a9f9, which
had lazily made the description less wrong without bothering to actually
fix it.