Simplify NativeHashedStorageHandler()
This avoids an opportunity for type(system) confusion, by always using
the full CompilerType instead of assuming that the type system hasn't
changed in CreateTupleType and GetCanononicalType(). Also the call to
GetCanonicalType() and IsTupleType was redundant.

Unfortunately I was not able to construct a reproducer for the type
confusion using a REPL-defined alias type, but I received a crashlog
that seems indicate that it's indeed possible.