[DependencyScanningCASFilesystem] Fix issue where the PP skipped range mappings are not set when doing CAS dep-scanning
Previously the skipped range mappings were only applied to the preprocessor's `PPSkipMappings` when getting the minimized info,
but the minimized info was getting cached for subsequent invocations while the preprocessor was clearing out its `PPSkipMappings`;
this resulted in the PP skipped range optimization only being active for the first invocation of a `DependencyScanningWorker`.

To fix this, do as `DependencyScanningFilesystem` and cache the range mappings and apply them when `openFileForRead()` is called.

This reduces CAS dep-scanning time for clang sources by ~ -17%.