[ObjCARC] Tolerate missing attachedcall op by ignoring the bundle.
In some situations (swiftc bitcode) we end up running the
objc-arc-contract pass twice, which is not supposed to occur.

This was fine until relatively recently, where the
clang.arc.attachedcall operand bundle support caused the second run to
crash: when emitting the calls, the pass was assuming that the bundle
always had an operand, which is not true in 5.6.

That will be true in the future, but that's a massive change.
For now, try to workaround the issue by ignoring the nullary bundles,
and not emitting the call.

The pass will get a little confused, but the semantics are preserved
by the presence of the intrinsic call later on (emitted by the first
run of the pass).  However, one difference is that we're now going
to emit a redundant inline-asm marker, since we're not able to
eliminate the (standalone) intrinsic anymore, as we don't remember
emitting it in the pass.  That costs us an extra nop, which is okay.

rdar://90366906