[lldb] Fix invalid use of mutating keyword in generated expressions
Prevent conditional logic from falling through to adding the mutating
keyword when evaluating Swift expressions inside of a a function whose
self type is a dynamic Self.

After determining the type of self, resolve instances of DynamicSelfType
to their self type. This allows consumers to properly determine when a
type is a class, for example, which is not possible when using the
dynamic self directly.