[ADT] Remove AtomicUniquePointer
AtomicUniquePointer is not friendly to use with compare_exchange
interface because double reference to the pointer is unavoidable so it
adds more burden for user to track life time than the utility it
provides.

In the current codebase, it is only being used to tracking lifetime of
owning objects in the HashMappedTrie. It is cleaner just to handle the
lifetime of pointers in the destructor/destroyImpl.