[cas] Add a clang-cas-test utility with option to dump compile-job cache keys
Usage:
    clang-cas-test -print-compile-job-cache-key [-cas PATH] <KEY>

This makes it easier to debug what went into a given compile-job cache
key and to compare it to another key by printing out the command-line,
filesystem, and other entries in a format that is amenable to diff.

Incidentally, factor the code for creating a cache key into the same
file as the printer.