[driver] Introduce `clang-cache` driver variant, intended to be used as a compiler launcher for compilation caching
`clang-cache` provides capability to use compilation caching without needing to modify any of the compiler arguments.
It is intended to be used as the compiler launcher along with using the same clang binary for compilation.

Environment variable `CLANG_CACHE_CAS_PATH` can be used to control the on-disk CAS path, otherwise the default location is used.