[driver] Introduce `clang-cache-build-session`
`clang-cache-build-session` can be used in combination with `clang-cache`
to start a build command that indicates a "build session". After configuring
to use `clang-cache` as the compiler launcher, you can use it like this:

```
% $TOOLCHAIN/usr/bin/clang-cache-build-session ninja
```

`clang-cache-build-session` accepts a command that triggers a build and,
while the command is running, all the clang invocations that are part of
that build will be sharing the same dependency scanning daemon.

This is accomplished by having `clang-cache-build-session` set
`CLANG_CACHE_BUILD_SESSION_ID=<pid>` as environment variable which is then
passed on to the new "depscan" flag `-fdepscan-share-identifier`.